Beta alpha gamma version for diploma

+++ gplot-core/include/Graphics/Mesh.hpp
@@ -12,13 +12,13 @@ namespace gplot::graphics
     {
     public:
 
-        Mesh(const std::vector<Vertex>& vertices, const std::vector<GLushort>& indices)
+        Mesh(const std::vector<Vertex>& vertices, const std::vector<GLuint>& indices)
             : Mesh(vertices.data(), vertices.size(), indices.data(), indices.size())
         {
         }
 
         template<size_t Vertices, size_t Indices>
-        Mesh(const std::array<Vertex, Vertices>& vertices, const std::array<GLushort, Indices>& indices)
+        Mesh(const std::array<Vertex, Vertices>& vertices, const std::array<GLuint, Indices>& indices)
             : Mesh(vertices.data(), vertices.size(), indices.data(), indices.size())
         {
         }
@@ -27,15 +27,15 @@ namespace gplot::graphics
 
     private:
 
-        Mesh(const Vertex* vert_data, size_t vert_size, const GLushort* ind_data, size_t ind_size)
+        Mesh(const Vertex* vert_data, size_t vert_size, const GLuint* ind_data, size_t ind_size)
             : m_count(ind_size)
             , m_buffer(GetDescriptor())
         {
             m_buffer.Resize<Vertex>(0, gplot::graphics::VertexBuffer::BufferData_t::eVertexData,vert_size);
             m_buffer.Update<Vertex>(0, gplot::graphics::VertexBuffer::BufferData_t::eVertexData, vert_size, vert_data);
 
-            m_buffer.Resize<GLushort>(0, gplot::graphics::VertexBuffer::BufferData_t::eIndicesData,ind_size);
-            m_buffer.Update<GLushort>(0, gplot::graphics::VertexBuffer::BufferData_t::eIndicesData,ind_size, ind_data);
+            m_buffer.Resize<GLuint>(0, gplot::graphics::VertexBuffer::BufferData_t::eIndicesData,ind_size);
+            m_buffer.Update<GLuint>(0, gplot::graphics::VertexBuffer::BufferData_t::eIndicesData,ind_size, ind_data);
         }
 
         static VertexBuffer::VertexBufferDescriptor GetDescriptor();
